Fix pattern n_entries calculation, which was broken when attempting to fix the warning incorrectly with long casts, which resulted in n_entries that were horribly inflated, possibly breaking code that didn't expect the change.

+++ src/it/readam.c
@@ -255,7 +255,7 @@ static int it_riff_am_process_pattern(IT_PATTERN *pattern, DUMBFILE *f,
         ++row;
     }
 
-    pattern->n_entries = (int)((long)entry - (long)pattern->entry);
+    pattern->n_entries = (int)(entry - pattern->entry);
     if (!pattern->n_entries)
         return -1;
